[UPDATED] changes to Type typeguard generation
