+++ src/api/generateIntersectionTypeGuard.ts
@@ -1,0 +1,14 @@
+import {isIntersectionTypeNode, TypeNode} from "typescript";
+import {generateKeywordGuard} from "../api";
+import {generateTypeReferenceGuard} from "../api";
+
+export function generateIntersectionTypeGuard(type: TypeNode, propertyName: string) {
+    const typeGuard: string[] = [];
+    if (!isIntersectionTypeNode(type)) return typeGuard;
+    if (!type.types) return typeGuard;
+    for (const member of type.types) {
+        typeGuard.push(...generateKeywordGuard(member, propertyName));
+        typeGuard.push(...generateTypeReferenceGuard(member, propertyName));
+    }
+    return typeGuard;
+}
