+++ src/api/index.ts
@@ -1,0 +1,8 @@
+export {generateInterfaceTypeGuard} from "./generateInterfaceTypeGuard";
+export {generateIntersectionTypeGuard} from "./generateIntersectionTypeGuard";
+export {generateKeywordGuard} from "./generateKeywordGuard";
+export {generatePropertyGuard} from "./generatePropertyGuard";
+export {generateSingleInterfaceTypeGuard} from "./generateSingleInterfaceTypeGuard";
+export {generateTypeReferenceGuard} from "./generateTypeReferenceGuard";
+export {generateUnionTypeGuard} from "./generateUnionTypeGuard";
+export {handleHeritageClauses} from "./handleHeritageClauses";
