+++ src/data.ts
@@ -5,6 +5,7 @@ export interface Person {
     unionReferenceType?: string | Point;
     literalType: number;
     referenceType: Point;
+    anyType?: any;
 }
 
 export interface Person2 extends Person {
@@ -22,34 +23,30 @@ export type Point = {
     x: number;
     y: number;
 };
-//TODO: add checks for Parenthesis types
-/*
+export type someType = "a" | null | undefined | boolean | number | string;
 export type PersonType = {
   name: string & number;
   extra: string | number;
 };
-//TODO: fix generator for & in types. Does not generate !value.hasOwnProperty('name') for 'extra' property
-export type Address = {
+//TODO
+//Intersection types
+//Parenthesis types
+
+export type AddressType = {
   street: string;
 } & PersonType;
-export type Point2 = {
+export type Point2Type = {
   x: number;
   y: number;
   z?: Point;
 } & Point;
-export type Point3 = {
+export type Point3Type = {
   z?: number;
   m?: string & Point;
-} & Point &
-  Point2;
+} & number;
 
-/!*
-export enum Color {
+/*export enum Color {
     Red = 'red',
     Green = 'green',
     Blue = 'blue',
-}
-//TODO: add support for | in types
-// | Point | boolean | null | undefined*!/
-export type someType = "a" | null | undefined | boolean | number | string;
-*/
+}*/
