+++ src/generator/generateConditionalTypeGuard.ts
@@ -1,5 +1,6 @@
 // Generate type guards for conditional types
-import { generateInterfaceTypeGuard } from "./generateInterfaceTypeGuard";
+
+import {generateInterfaceTypeGuard} from "../api/generateInterfaceTypeGuard";
 
 export function generateConditionalTypeGuard(
   typeName: string,
