+++ src/generator/generateTypeGuard.ts
@@ -1,7 +1,7 @@
 // Generate a type guard for a given interface or type
-import { generateInterfaceTypeGuard } from "./generateInterfaceTypeGuard";
 import { generateEnumTypeGuard } from "./generateEnumTypeGuard";
 import { isObject } from "../utils";
+import {generateInterfaceTypeGuard} from "../api/generateInterfaceTypeGuard";
 
 export function generateTypeGuard(typeName: string, definition: any) {
   /*let typeGuard = `function is${typeName}(value: any): value is ${typeName} {\n`;
