+++ src/index.ts
@@ -26,7 +26,9 @@ import {
   getMembersFromTypeAlias,
 } from "./utils";
 import { generateTypeGuards, generateUnionTypeGuard } from "./generator";
-import {generateInterfaceTypeGuard} from "./generator/generateInterfaceTypeGuard";
+
+import {generateInterfaceTypeGuard} from "./api";
+import {generateTypeTypeGuard} from "./api/generateTypeTypeGuard";
 
 type ObjectsType = {
   interfaces: ts.InterfaceDeclaration[];
@@ -157,4 +159,4 @@ function loadConfig() {
 //Implementation
 const { interfaces, types, enums } = readObjects(loadConfig());
 deleteFileIfExists("out/typeguards.ts");
-generateTypeGuardsFile(`${generateTypeGuards(types)}\n${generateInterfaceTypeGuard(interfaces)}`);
+generateTypeGuardsFile(`${generateTypeTypeGuard(types)}\n${generateInterfaceTypeGuard(interfaces)}`);
